requirements.txt
Django==4.2.24
djangorestframework==3.16.1
Pillow==11.3.0
django-cors-headers==4.8.0
django-extensions==4.1
python-decouple==3.8
django-allauth==65.11.2
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
psycopg2-binary==2.9.10
celery==5.5.3
redis==6.4.0
django-celery-beat==2.8.1
django-environ==0.11.2
django-ckeditor==6.7.0
django-mptt==0.16.0
django-filter==24.3
django-import-export==4.1.1
plotly==5.22.0
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
django-widget-tweaks==1.5.0
jdatetime==5.2.0
django-jalali-date==2.0.0
